% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{annot_tsv_path}
\alias{annot_tsv_path}
\title{Get Path to annot-tsv Executable}
\usage{
annot_tsv_path()
}
\value{
A character string containing the path to the annot-tsv executable.
}
\description{
Returns the path to the bundled annot-tsv executable.
}
\examples{
annot_tsv_path()

}
