% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{bcftools_lib_dir}
\alias{bcftools_lib_dir}
\title{Get bcftools Library Directory}
\usage{
bcftools_lib_dir()
}
\value{
A character string containing the path to the bcftools lib directory.
}
\description{
Returns the path to the bcftools library files for use in linking.
}
\details{
This directory contains \code{libbcftools.a} (static) and \code{libbcftools.so}
(shared) libraries.
}
\examples{
bcftools_lib_dir()

}
