% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{bcftools_path}
\alias{bcftools_path}
\title{Get Path to bcftools Executable}
\usage{
bcftools_path()
}
\value{
A character string containing the path to the bcftools executable.
}
\description{
Returns the path to the bundled bcftools executable.
}
\examples{
bcftools_path()

}
