% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{bcftools_plugins_dir}
\alias{bcftools_plugins_dir}
\title{Get Path to bcftools Plugins Directory}
\usage{
bcftools_plugins_dir()
}
\value{
A character string containing the path to the bcftools plugins
directory.
}
\description{
Returns the path to the directory containing bcftools plugins.
}
\examples{
bcftools_plugins_dir()

}
