% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{bgzip_path}
\alias{bgzip_path}
\title{Get Path to bgzip Executable}
\usage{
bgzip_path()
}
\value{
A character string containing the path to the bgzip executable.
}
\description{
Returns the path to the bundled bgzip executable.
}
\examples{
bgzip_path()

}
