% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducklake.R
\name{ducklake_current_snapshot}
\alias{ducklake_current_snapshot}
\title{Get current snapshot ID}
\usage{
ducklake_current_snapshot(con, catalog = "lake")
}
\arguments{
\item{con}{DuckDB connection with DuckLake attached.}

\item{catalog}{DuckLake catalog name.}
}
\value{
Integer snapshot ID.
}
\description{
Get current snapshot ID
}
