% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducklake.R
\name{ducklake_load}
\alias{ducklake_load}
\title{Load the DuckLake extension}
\usage{
ducklake_load(con, install = TRUE)
}
\arguments{
\item{con}{A DuckDB connection.}

\item{install}{Logical, attempt \verb{INSTALL ducklake} before loading. Defaults to TRUE.}
}
\value{
The connection (invisibly).
}
\description{
Load the DuckLake extension
}
