% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducklake.R
\name{ducklake_options}
\alias{ducklake_options}
\title{Get DuckLake configuration options}
\usage{
ducklake_options(con, catalog = "lake")
}
\arguments{
\item{con}{DuckDB connection with DuckLake attached.}

\item{catalog}{DuckLake catalog name.}
}
\value{
Data frame with current options.
}
\description{
Get DuckLake configuration options
}
