% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{htsfile_path}
\alias{htsfile_path}
\title{Get Path to htsfile Executable}
\usage{
htsfile_path()
}
\value{
A character string containing the path to the htsfile executable.
}
\description{
Returns the path to the bundled htsfile executable for identifying file
formats.
}
\examples{
htsfile_path()

}
