% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{htslib_bin_dir}
\alias{htslib_bin_dir}
\title{Get Path to htslib Binary Directory}
\usage{
htslib_bin_dir()
}
\value{
A character string containing the path to the htslib bin directory.
}
\description{
Returns the path to the directory containing htslib executables.
}
\details{
The directory contains the following tools:
\itemize{
\item \code{annot-tsv} - Annotate TSV files
\item \code{bgzip} - Block gzip compression
\item \code{htsfile} - Identify file format
\item \code{ref-cache} - Reference sequence cache management
\item \code{tabix} - Index and query TAB-delimited files
}
}
\examples{
htslib_bin_dir()

}
