% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{htslib_features}
\alias{htslib_features}
\title{Get htslib Features Bitfield}
\usage{
htslib_features()
}
\value{
An integer representing the feature bitfield.
}
\description{
Returns the raw bitfield of enabled features in htslib.
}
\examples{
htslib_features()

}
