% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\docType{data}
\name{htslib_has_feature}
\alias{htslib_has_feature}
\alias{HTS_FEATURE_CONFIGURE}
\alias{HTS_FEATURE_PLUGINS}
\alias{HTS_FEATURE_LIBCURL}
\alias{HTS_FEATURE_S3}
\alias{HTS_FEATURE_GCS}
\alias{HTS_FEATURE_LIBDEFLATE}
\alias{HTS_FEATURE_LZMA}
\alias{HTS_FEATURE_BZIP2}
\alias{HTS_FEATURE_HTSCODECS}
\title{Check for a Specific htslib Feature}
\format{
An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
htslib_has_feature(feature_id)

HTS_FEATURE_CONFIGURE

HTS_FEATURE_PLUGINS

HTS_FEATURE_LIBCURL

HTS_FEATURE_S3

HTS_FEATURE_GCS

HTS_FEATURE_LIBDEFLATE

HTS_FEATURE_LZMA

HTS_FEATURE_BZIP2

HTS_FEATURE_HTSCODECS
}
\arguments{
\item{feature_id}{An integer feature ID. Use one of the \verb{HTS_FEATURE_*}
constants.}
}
\value{
A logical value indicating if the feature is enabled.
}
\description{
Checks if a specific feature is enabled in the bundled htslib library.
}
\examples{
# Check for libcurl support (feature ID 1024)
htslib_has_feature(1024L)

}
\keyword{datasets}
