% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_duckdb.R
\name{parquet_to_vcf_tidy}
\alias{parquet_to_vcf_tidy}
\title{Convert tidy format Parquet back to VCF}
\usage{
parquet_to_vcf_tidy(input_file, output_file, header, index, con, meta)
}
\description{
Convert tidy format Parquet back to VCF
}
\keyword{internal}
