% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{rducks_duckdb_types}
\alias{rducks_duckdb_types}
\title{Convert Rducks type descriptors to DuckDB SQL types}
\usage{
rducks_duckdb_types(x)
}
\arguments{
\item{x}{Character scalar tokens, \code{rducks_type} descriptors, or a list of descriptors.}
}
\value{
Character vector of DuckDB SQL type names.
}
\description{
Convert Rducks type descriptors to DuckDB SQL types
}
\examples{
rducks_duckdb_types(INTEGER)
rducks_duckdb_types(c(INTEGER, DOUBLE, VARCHAR))
rducks_duckdb_types(STRUCT(a = INTEGER, b = VARCHAR))
}
