% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exotic_classes.R
\name{rducks_union}
\alias{rducks_union}
\title{Construct DuckDB UNION values}
\usage{
rducks_union(tag, value)
}
\arguments{
\item{tag}{Character scalar union member name.}

\item{value}{R value for that member.}
}
\value{
Object of class \code{rducks_union}.
}
\description{
\code{rducks_union()} represents one tagged union value. The tag should match a
DuckDB union member name; \code{value} is the corresponding R value.
}
\examples{
rducks_union("num", 42L)
rducks_union("str", "hello")
}
