% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{setup_hts_env}
\alias{setup_hts_env}
\title{Setup Environment for Remote File Access}
\usage{
setup_hts_env()
}
\value{
Invisibly returns the previous value of \code{HTS_PATH} (or \code{NA} if unset).
}
\description{
Sets the \code{HTS_PATH} environment variable to point to the bundled htslib
plugins directory. This is required for S3, GCS, and other remote file
access via libcurl.
}
\details{
Call this function before using bcftools/htslib tools with remote URLs
(s3://, gs://, http://, etc.). The function sets \code{HTS_PATH} to the package's
plugin directory so htslib can find \code{hfile_libcurl.so} and \code{hfile_gcs.so}.
}
\examples{
setup_hts_env()
# Now bcftools can access S3 URLs

}
