% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{tabix_path}
\alias{tabix_path}
\title{Get Path to tabix Executable}
\usage{
tabix_path()
}
\value{
A character string containing the path to the tabix executable.
}
\description{
Returns the path to the bundled tabix executable.
}
\examples{
tabix_path()

}
