% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_arrow.R
\name{vcf_arrow_schema}
\alias{vcf_arrow_schema}
\title{Get the Arrow schema for a VCF file}
\usage{
vcf_arrow_schema(filename)
}
\arguments{
\item{filename}{Path to VCF or BCF file}
}
\value{
A nanoarrow_schema object
}
\description{
Reads the header of a VCF/BCF file and returns the corresponding
Arrow schema.
}
