% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_duckdb.R
\name{vcf_open_duckdb_parallel_view}
\alias{vcf_open_duckdb_parallel_view}
\title{Internal: Create parallel VIEW using UNION ALL of per-contig bcf_read calls}
\usage{
vcf_open_duckdb_parallel_view(
  con,
  file,
  table_name,
  columns,
  tidy_format,
  threads
)
}
\arguments{
\item{con}{DuckDB connection}

\item{file}{VCF file path}

\item{table_name}{Target view name}

\item{columns}{Columns to select}

\item{tidy_format}{Whether to use tidy format}

\item{threads}{Number of threads (used to limit contigs)}
}
\value{
NULL (views don't have row counts)
}
\description{
Creates a VIEW that unions bcf_read() calls for each contig. DuckDB can
parallelize execution at query time.
}
\keyword{internal}
