% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep_parser.R
\name{vep_has_annotation}
\alias{vep_has_annotation}
\title{Check if VCF has VEP-style annotations}
\usage{
vep_has_annotation(filename)
}
\arguments{
\item{filename}{Path to VCF/BCF file}
}
\value{
Logical indicating presence of CSQ, BCSQ, or ANN
}
\description{
Check if VCF has VEP-style annotations
}
\examples{
\dontrun{
if (vep_has_annotation("file.vcf.gz")) {
  schema <- vep_get_schema("file.vcf.gz")
}
}

}
