% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep_parser.R
\name{vep_list_fields}
\alias{vep_list_fields}
\title{List VEP annotation fields in a VCF file}
\usage{
vep_list_fields(filename)
}
\arguments{
\item{filename}{Path to VCF/BCF file}
}
\value{
Invisibly returns the schema data frame
}
\description{
Convenience function to display available VEP fields and their types.
}
\examples{
\dontrun{
vep_list_fields("annotated.vcf.gz")
# VEP Annotation Tag: CSQ
# Fields (78 total):
#   1. Allele (String)
#   2. Consequence (String, list)
#   3. IMPACT (String)
#   ...
}

}
